+++ ludwig/schema/decoders/__init__.py
@@ -1,3 +1,4 @@
 # Register all decoders
 import ludwig.schema.decoders.base
+import ludwig.schema.decoders.llm_decoders  # noqa
 import ludwig.schema.decoders.sequence_decoders  # noqa
